add global network switch
