+++ src/components/Template/components/Navbar/Navbar.tsx
@@ -1,12 +1,11 @@
-import { MouseEvent, useCallback } from 'react';
+import React, { MouseEvent, useCallback, useState } from 'react';
 import { Link } from 'react-router-dom';
 import classNames from 'classnames';
-
 import { MultiversXLogo } from 'assets/img/MultiversXLogo';
-
 import styles from './styles.module.scss';
-
 import type { NavbarPropsType } from './types';
+import { useChain } from '../../../../hooks/useChain';
+import { Environment } from '../Environment';
 
 /*
  * Handle the component declaration.
@@ -14,6 +13,8 @@ import type { NavbarPropsType } from './types';
 
 export const Navbar = (props: NavbarPropsType) => {
   const { setToggleMenu, toggleMenu } = props;
+
+  const { chain } = useChain();
 
   /*
    * On menu triggering, update the passed along state to the opposite of the current boolean value.
@@ -41,6 +42,7 @@ export const Navbar = (props: NavbarPropsType) => {
         </Link>
 
         <div className={styles.right}>
+          <Environment chain={chain} />
           <div
             onClick={onMenuTrigger}
             data-testid='navbar-burger'
